Update interlink converter model and others
The interlink converter is more likely to be unstable caused by the numerical calculation. Therefore, the discrete damping resistor of it is reduced in the simulink model.

Due to the matlab version compatibility, only the 2020a can be used for running the toolbox.

The HVDC test system is added in hybrid example system.